+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Continue add item if updateOrderFormMarketingData fails and just log its error.
 
 ## [2.123.2] - 2020-05-18
 ### Fixed